src/requirements.txt
asgiref==3.11.1
Django==5.2
django-htmx==1.27.0
django-taggit==6.1.0
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.1
RapidFuzz==3.14.3
sqlparse==0.5.5
thefuzz==0.22.1
typing_extensions==4.15.0
